main: Update code 20260621

+++ SurveyReportRE.Models/Models/Migration/Business/Data/SLA.cs
@@ -12,7 +12,7 @@ public class SLA : BaseModel
     [MaxLength(50)]
     public string? Code { get; set; }
     public long? Value { get; set; }    
-    public long? DecimalValue { get; set; }     
+    public double? DecimalValue { get; set; }     
     public DateTime? FromDate { get; set; } 
     public DateTime? ToDate { get; set; } 
     public TimeSpan? Duration { get; set; }
